% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basketball.draft.development}
\alias{basketball.draft.development}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 200 rows and 9 variables:
\describe{
\item{player_id}{Player identifier.}
\item{player_height}{Player height, in cm.}
\item{player_position}{Player position: Guard, Forward or Center.}
\item{ncaa_48pts}{Player points per game in the NCAA.}
\item{ncaa_assist_pct}{Player assist percentage in the NCAA per 100 units.}
\item{ncaa_turnover_pct}{Player turnover percentage in the NCAA per 100 units.}
\item{ncaa_usage_rate}{Player usage rate in the NCAA.}
\item{nba_40pts}{Player points per game in the NBA.}
\item{nba_winshare}{Player WinShare in the NBA.}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
basketball.draft.development
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
