% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb.draft.development}
\alias{bb.draft.development}
\title{NBA MOCK draft.}
\format{
A data frame with 200 rows and 9 variables:
\describe{
\item{player_id}{Player identifier.}
\item{player_height}{Player height, in cm.}
\item{player_position}{Player position: Guard, Forward or Center.}
\item{NCAA_40pts}{Player points per game in the NCAA.}
\item{NCAA_assist_pct}{Player assist percentage in the NCAA per 100 units.}
\item{NCAA_turnover_pct}{Player turnover percentage in the NCAA per 100 units.}
\item{NCAA_usage_rate}{Player usage rate in the NCAA.}
\item{NBA_48pts}{Player points per game in the NBA.}
\item{NBA_winshare}{Player WinShare in the NBA.}
}
}
\source{
\url{}
}
\usage{
bb.draft.development
}
\description{
A dataset containing players that played in college and in the NBA.
}
\keyword{datasets}
