% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb.epl.matches1718}
\alias{fb.epl.matches1718}
\title{English premier league statistics.}
\format{
A data frame with 380 rows:
\describe{
\item{match_id}{Unique identifier of each match, containing league abbreviation (country), abbreviated team names and season.}
\item{home}{Name of the home team.}
\item{away}{Name of the away team.}
\item{home_goals}{Number of goals scored in the match by the home team.}
\item{away_goals}{Number of goals scored in the match by the away team.}
\item{sum_pass_length}{Total distance covered by the ball, summed across all passes completed in the match by both teams in m.}
\item{mn_pass_length}{Mean distance covered by the ball per pass across all passes completed in the match by both teams in m.}
\item{goals}{Total number of goals scored in the match by both teams.}
}
}
\source{
Pappalardo, L., Cintia, P., Rossi, A., Massucco, E., Ferragina, P., Pedreschi, D., & Giannotti, F. (2019). A public data set of spatio-temporal match events in soccer competitions. Scientific data, 6(1), 1-15.
}
\usage{
fb.epl.matches1718
}
\description{
All 380 matches from the English Premier League including home and away team,
home goals, away goals, total number of goals and total pass distance per match
}
\keyword{datasets}
