% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb.europe.marketvalue1718}
\alias{fb.europe.marketvalue1718}
\title{2017-2018 Player Market values in European Football.}
\format{
A data frame with 2372 rows:
\describe{
\item{row_id}{Row number.}
\item{player_name}{Name of the player.}
\item{marketval}{Market Value of the player.}
\item{position_tm}{Position playing in current team.}
\item{Competition}{Football competition currently playing in.}
\item{team}{Player's team.}
\item{Name.Wyscout}{Player name as represented in Wyscout data sources.}
}
}
\source{
Transfermarkt.de
}
\usage{
fb.europe.marketvalue1718
}
\description{
2017-2018 Player Market values in European Football.
}
\keyword{datasets}
