% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb.europe.playerxg1718}
\alias{fb.europe.playerxg1718}
\title{2017-2018 Player Statistics on Expected Goals and Expected Assists.}
\format{
A data frame with 2372 rows:
\describe{
\item{player}{Name of the player.}

\item{xg_Aerial}{Aerial Expected goals generated by the player.}
\item{xg_Cross}{Expected goals by crossing generated by the player.}
\item{xg_None}{Expected goals.}
\item{xg_Pass}{Expected goals by passing generated by the player.}
\item{xg_Dribble}{Expected goals by dribbling generated by the player.}
\item{xg_LooseBall}{Expected goals by loosing the ball generated by the player.}
\item{xg_Chipped}{Expected goals by chipping the ball generated by the player.}
\item{xg_Throughball}{Expected goals by through passes generated by the player.}
\item{xg_LayOff}{Expected goals by layoff generated by the player.}

\item{n_shots_Aerial}{Aerial Number of shots generated by the player.}
\item{n_shots_Cross}{Number of shots by crossing generated by the player.}
\item{n_shots_None}{Number of shots.}
\item{n_shots_Pass}{Number of shots by passing generated by the player.}
\item{n_shots_Dribble}{Number of shots by dribbling generated by the player.}
\item{n_shots_LooseBall}{Number of shots by loosing the ball generated by the player.}
\item{n_shots_Chipped}{Number of shots by chipping the ball generated by the player.}
\item{n_shots_Throughball}{Number of shots by through passes generated by the player.}
\item{n_shots_LayOff}{Number of shots by layoff generated by the player.}

\item{xg_foot}{Expected goals by actions with the foot generated by the player.}
\item{xg_headother}{Expected goals by actions with the head or other parts generated by the player.}
\item{n_shots_foot}{Number of shots by actions with the foot generated by the player.}
\item{n_shots_headother}{Number of shots by actions with the head or other parts generated by the player.}

\item{xa_Aerial}{Aerial Expected assists generated by the player.}
\item{xa_Cross}{Expected assists by crossing generated by the player.}
\item{xa_None}{Expected assists.}
\item{xa_Pass}{Expected assists by passing generated by the player.}
\item{xa_Dribble}{Expected assists by dribbling generated by the player.}
\item{xa_LooseBall}{Expected assists by loosing the ball generated by the player.}
\item{xa_Chipped}{Expected assists by chipping the ball generated by the player.}
\item{xa_Throughball}{Expected assists by through passes generated by the player.}
\item{xa_LayOff}{Expected assists by layoff generated by the player.}

\item{n_assists_Aerial}{Aerial Number of assists generated by the player.}
\item{n_assists_Cross}{Number of assists by crossing generated by the player.}
\item{n_assists_None}{Number of assists.}
\item{n_assists_Pass}{Number of assists by passing generated by the player.}
\item{n_assists_Dribble}{Number of assists by dribbling generated by the player.}
\item{n_assists_LooseBall}{Number of assists by loosing the ball generated by the player.}
\item{n_assists_Chipped}{Number of assists by chipping the ball generated by the player.}
\item{n_assists_Throughball}{Number of assists by through passes generated by the player.}
\item{n_assists_LayOff}{Number of assists by layoff generated by the player.}

\item{xa_foot}{Expected assists by actions with the foot generated by the player.}
\item{xa_headother}{Expected assists by actions with the head or other parts generated by the player.}
\item{n_assists_foot}{Number of assists by actions with the foot generated by the player.}
\item{n_assists_headother}{Number of assists by actions with the head or other parts generated by the player.}

\item{totxg}{Total Expected Goals generated by the player.}
\item{totxa}{Total Expected Assists generated by the player.}

}
}
\source{
Understat data
}
\usage{
fb.europe.playerxg1718
}
\description{
2017-2018 Player Statistics on Expected Goals and Expected Assists.
}
\keyword{datasets}
