% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb.europe.xg1622}
\alias{fb.europe.xg1622}
\title{Expected Goals in European Leagues from 2016 until 2022.}
\format{
A data frame with 262951 rows:
\describe{
\item{league}{Name of the competition out of the top five European leagues.}
\item{id}{Unique ID of the respective shot.}
\item{minute}{Minute of play in which shot was taken.}
\item{result}{Outcome of the shot as either Goal, MissedShots, BlockedShot, SavedShot or ShotOnPost.}
\item{shot_x}{X-coordinate of the shot location in m on a pitch where x axis is parallel to the sidelines, both teams are considered playing from left ro right and the pitch has dimensions 105x68m.}
\item{shot_y}{Y-coordinate of the shot location in m on a pitch where y axis is parallel to the goalline, both teams are considered playing from left ro right and the pitch has dimensions 105x68m.}
\item{xG}{Expected Goals-Value of the Shot as given on the understat.com website.}
\item{player}{Name of the player taking the shot.}
\item{h_a}{Indicator of whether the team of the player taking the shot is playing home (h) or away (a).}
\item{situation}{Situation from which the shot was taken as either OpenPlay, FromCorner, Penalty, SetPiece or DirectFreekick.}
\item{season}{Starting year of the respective season.}
\item{shotType}{Indicator of whether the shot was taken with RightFoot, LeftFoot or Head.}
\item{match_id}{Unique ID of the respective match.}
\item{home_team}{Name of the home team.}
\item{away_team}{Name of the away team.}
\item{home_goals}{Number of goals scored by the home team in the whole match.}
\item{away_goals}{Number of goals scored by the away team in the whole match.}
\item{date}{Kickoff date and time of the match.}
\item{player_assisted}{Name of the player giving the assist for the shot.}
\item{lastAction}{Type of the action preceding the shot}
\item{outcome}{Outcome of the Shot as a binary variable, Goal or NoGoal.}
}
}
\source{
\url{https://understat.com/}
}
\usage{
fb.europe.xg1622
}
\description{
Including information on every shot, such as player,
X-/Y-coordinates, XG value.
}
\keyword{datasets}
