% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb.metrica.trackingH1}
\alias{fb.metrica.trackingH1}
\title{Detailed tracking data from a football game.}
\format{
A data frame with 141,158 rows:
\describe{
\item{Period}{Period of the game.}
\item{Frame}{Snapshot of the game.}
\item{Time_s}{Time in the game.}
\item{Player1x}{X coordinates of player 1.}
\item{Player1y}{Y coordinates of player 1.}
\item{Player2x}{X coordinates of player 2.}
\item{Player2y}{Y coordinates of player 2.}
\item{Player3x}{X coordinates of player 3.}
\item{Player3y}{Y coordinates of player 3.}
\item{Player4x}{X coordinates of player 4.}
\item{Player4y}{Y coordinates of player 4.}
\item{Player5x}{X coordinates of player 5.}
\item{Player5y}{Y coordinates of player 5.}
\item{Player6x}{X coordinates of player 6.}
\item{Player6y}{Y coordinates of player 6.}
\item{Player7x}{X coordinates of player 7.}
\item{Player7y}{Y coordinates of player 7.}
\item{Player8x}{X coordinates of player 8.}
\item{Player8y}{Y coordinates of player 8.}
\item{Player9x}{X coordinates of player 9.}
\item{Player9y}{Y coordinates of player 9.}
\item{Player10x}{X coordinates of player 10.}
\item{Player10y}{Y coordinates of player 10.}
\item{Player11x}{X coordinates of player 11.}
\item{Player11y}{Y coordinates of player 11.}
\item{Player12x}{X coordinates of player 12.}
\item{Player12y}{Y coordinates of player 12.}
\item{Player13x}{X coordinates of player 13.}
\item{Player13y}{Y coordinates of player 13.}
\item{Player14x}{X coordinates of player 14.}
\item{Player14y}{Y coordinates of player 14.}
\item{Ballx}{.}
\item{Bally}{.}
}
}
\source{
Metrica Sports https://github.com/metrica-sports/sample-data/tree/master/data
}
\usage{
fb.metrica.trackingH1
}
\description{
Detailed tracking data from a football game.
}
\keyword{datasets}
