% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t.atp.eventsfinal19}
\alias{t.atp.eventsfinal19}
\title{Detailed event data from Tennis pro matches.}
\format{
A data frame with 852 rows:
\describe{
\item{X}{Id of the row.}
\item{rallyid}{Rally identifier.}
\item{frameid}{Game score.}
\item{strokeid}{Stroke identifier.}
\item{hitter}{Hitter name.}
\item{receiver}{Receiver name.}
\item{isserve}{Flag indicated wether the action was the serve.}
\item{serve}{Serve.}
\item{type}{Type of the hit.}
\item{stroke}{Storke of the hit.}
\item{hitter_x}{Coordinate X of the hitter.}
\item{hitter_y}{Coordinate Y of the hitter.}
\item{receiver_x}{Coordinate X of the receiver.}
\item{receiver_y}{Coordinate Y of the receiver.}
\item{time}{Timestamp of the action.}
}
}
\source{
Kaggle
}
\usage{
t.atp.eventsfinal19
}
\description{
ATP final 2019: Djokovic vs. Nadal
}
\keyword{datasets}
