% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t.pbp.all}
\alias{t.pbp.all}
\title{Detailed shot-by-shot data from Tennis pro matches.}
\format{
A data frame with 60K rows:
\describe{
\item{Server}{The player serving.}
\item{Sets}{Set score.}
\item{Games}{Game score.}
\item{Points}{Points in the set}
\item{Description}{Detailed rally description, standard definition in english separated by semicolon.}
}
}
\source{
\url{http://www.tennisabstract.com/charting/meta.html}
}
\usage{
t.pbp.all
}
\description{
User added detailed information of every shot in tennis matches
for all Grand Slams from 2011 until 2022. Data has been obtained thanks
the work of Jeff Sackmann under the Match Charting Project.
}
\keyword{datasets}
